[android][sdk] Move Android7RootCertificateWorkaround into sdk package
