+++ android/sdk/src/main/res/raw/globalsignr4
