+++ android/sdk/src/main/res/raw/gtsrootr3
